requirements.txt
# requirements.txt
streamlit
pandas
numpy
nltk
google-genai